core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["DimaPekutko <dkefir03@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asgiref= ">=3.5.2"
Django = ">=4.0.6"
djangorestframework=">=3.13.1"
python-dotenv=">=0.20.0"
pytz=">=2022.1"
sqlparse=">=0.4.2"
pyjwt=">=2.3.0"
pytest-django=">=4.5.2"
celery=">=5.2.7"
boto3=">=1.24.30"
pika = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
